requirements.txt
pymupdf==1.24.13
Pillow==10.4.0
watchdog==5.0.3
